feat: allow filtering of actors & credits during grid gen
This is a pretty big change to the grid-generation algorithm and adds
some complexity to it. I spent a while debugging this. It's becoming
a bit messy, and I think could benefit from a refactor/clean up.

Most importantly, this commit adds 2 parameters to
`getValidAcrossAndDown`, `actorConditions` and `creditConditions`. Each
of these is a list of functions which return `true` or `false` for
a given actor/credit. If any condition in the list returns `true` for
the actor/credit being considered, that actor/credit is skipped. This
allows for more fine-grain contol over the actors and credits chosen for
the grid. For example, this allows the grid-gen algorithm to skip over
"Saturday Night Live" as a valid credit. While playing the game, SNL
will still be a valid answer, but it's guaranteed that SNL isn't the
**only** answer.

This commit also adds `genre_ids` to credits, which allows us to filter
on genre.