feat: totally refactor grid-gen algo, using strings as IDs
This commit includes a totally revamped algorithm for generating the
grid. It doesn't contain any filtering of actors or credits like the
previous one did, nor does it have the 'random' feature, but it's
abstracted away from the concepts of actors and credits. Because of this
commit, it would be very easy to swap the "roles" of actors and credits
in the grid, putting movies on the axes and finding actors as answers.

This algorithm uses strings to identify entities in the grid. This means
that the IDs of the actors and of the movies have been converted to
strings, even in the graph file in which they are stored. This was done
to generalize the algorithm. The final grid, when exported, still
contains number IDs, so nothing should change beyond the point of
generating the grid.