fix: use all credits (even filtered ones) in output grid
This was a huge effort. Because I was filtering a shallow copy of the
graph, I was modifying the original one as well. This made it impossible
to output a grid that contained all credits; the output grid would only
contain the credits that remained after filtering.

Implementing and using a deep copy of the graph proved more difficult
than expected. There were all sorts of referencing issues that came up,
and I'm not really sure how to learn from this except for that I'll now
be thinking about this as a potential source of bugs in the future. If
I had had tests in place, maybe some of these issues would have been
resolved. I'm unconvinced of that because I spent a **long** time
debugging, and the issues weren't apparent at all; the tests are only as
good as the person who writes them, and I'm not sure that my tests would
have been thorough enough to catch the kinds of bugs I encountered.

Anyway, now it's all working, so I can pre-filter to my heart's content.