fix: render correct height on mobile
On mobile, the height of the URL bar and the navigation button bar were
being taken into account, which makes the page scrollable and the bottom
of the page is cut off before any scrolling is done. dvh units seem to
resolve this problem, and I haven't noticed any other side effects.