feat: consider last air date for release-date categories
TV shows air over a period of time, and every release should be
considered when evaluating whether or not it was released "in the 21st
century," for example. This PR makes changes to save the last_air_date
field to credit extra info, and then uses that information to assess
whether a TV show "released" in the 21st century or not.