refactor: move to port/adapter arch for reading/writing graph
In order to simplify the code, I created a GraphHandler abstract base
class that sets up the methods required for reading and writing the
graph. This will be useful once we start implementing the database
functionality. This way, all we have to do to switch between them (or
others in the future) is to pass a different instance of GraphHandler.