feat: store categories upon data-store population
Previously, adding categories into the game was much more complicated.
I was adding them in once we had the actor/credit graph created, but
this lead to many issues related to storing the categories in the
database, in addition to more complicated code that distinguished
between actors and categories, even though they're really the same.

These changes allow us to store categories in the database, which in
turn let us store grids (which contain categories) in the database. Now
that grids can be stored in the database properly, we can start storing
answers and scores which correspond to a grid, which will let us achieve
our ultimate goal of storing and displaying stats.