chore: remove FileGraphHandler
When I started working on this project, I was using files to handle
storing the actors, credits, and extra credit info. When I moved to
using a database, I moved to a ports/adapters architecture, where either
the files or the database could be used. Since I'm not using the files
anymore, and additional functionality was being added to the database
code, I decided to just delete the FileGraphHandler. Now, there's only
1 adapter for this port, but adapters may be added in the future. That's
  why, at this point, you must specify graphMode as "db" in the Lambdas.
  In the future, I may add other adapters.