feat: allow saving of individual guesses
This is a bit of a refactor on top of the new feature. Firstly, we
rename 'Answer' to 'Guess', to make it clear that it may not be correct.
I renamed some Lambda functions as well because I thought their names
needed clarification, but who knows with naming...

This commit allows the frontend to send guesses as soon as they are
guessed, which will allow us to display data about that guess on the
fly. For example, if you guess a certain movie, the UI can immediately
show you (thanks to the response from the API) that the guessed movie
was used in (e.g.,) 10% of all correct guesses in that square. This was
previously not the case, as the setup forced us to wait until the game
was complete to save any guesses.