refactor: record stats for guesses immediately
Previously, guessing would only store a guess into the DB, and return
the guess's ID. At the end of the game, all guess IDs would be used to
update the score and link guesses to scores. I realized I could update
everything in real time, so now it works this way: upon guessing for the
first time, a score is created. Every subsequent guess has that new
score's ID attached to it, so that the guess can be entered in to the DB
with the new score on the fly.