refactor: store grid answers/credits as sets/objs respectively
This refactoring was mainly done to resolve a bug where the game summary
would take longer to load when there were more answers in the grid. For
example, if there were two very broad categories crossing in the grid
that resulted in a square with multiple thousands of answers, the game
summary would take a second to appear. The reason is that it was running
in O(credits^2) time.