refactor: rename 'connections' to 'links'; add interfaces for 'graph data'
I realized that a graph has AxisEntities and Connections, but within
each of those, there are `connections`, which link one to another.
I took the time to rename the inner `connections` to `links` to avoid
this overlapping of names.

Additionally, I added interfaces for `GraphData`, which will contain the
information required to actually put together a linked graph.