fix: give all credits unique IDs in graph
Credits need all to have unique IDs for the generic grid-generation
algorithm to work. The actual "id" field on `Credit`s is a number, but
movies and tv shows can have the same numeric ID. In order to fix this,
this commit adds a method to create a unique "ID" for a credit, which
looks like `movie-123` or `tv-456`.