feat: add store-data methods (untested)
It's a little bit difficult to test these methods without writing at
least one DataScraper, so I'm just leaving this untested for now and
will come back to it once the TMDBDataScraper is implemented.

One important fix in this commit is that we use the "movie-123" format
of ID everywhere *except* in the database. This conforms to the data
format expected everywhere else. Only in the database is this separation
made (and even there we could probably have done without it, but it's
a bit late for that... a future refactor could change the database
schema to allow for string IDs for credits that we set as "movie-123"
ourselves. Probably not even that much work honestly.)